refactor: remove the `resetter.rs`
